Fix synthesis problems with phase correction step
The output of the OS PFB in hardware was not as expected because the phase
correction step was not being syntehsized correctly. In fact, based on the output
it appeared that the shift was not being applied at all.

This was fixed by modifying the access patterns to the filter_out variable and
modifying how to determine the shift state. The filter_out are now accessed
sequentially between the FIR output and the FFT input.